Add 1..* data generator
Also deal with class imbalance in a better way. Instead of taking away random
points within a shot, we remove entire shots. This seems hugely important as a
key component of being able to discriminate between a fast shot and a slow shot
is how spaced out the locations are